requirements.txt
numpy>=1.21.0
Pillow>=8.2.0
matplotlib>=3.7.2